More changes to transition icecap to the New Way for runtime-manager
