+++ icecap-runtime-manager/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "icecap-runtime-manager"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+icecap-qemu = []
+
+[dependencies]
+anyhow = "1"
+bincode = { version = "1.2.1", default-features = false }
+icecap-core = { path = "../icecap/crates/framework/base/icecap-core" }
+icecap-rpc = { path = "../icecap/crates/framework/base/icecap-rpc" }
+icecap-ring-buffer = { path = "../icecap/crates/framework/base/icecap-ring-buffer" }
+icecap-start-generic = { path = "../icecap/crates/framework/base/icecap-start/generic" }
+icecap-std-external = { path = "../icecap/crates/framework/std-support/icecap-std-external" }
+libm = { version = "0.2" }
+psa-attestation = { path = "../psa-attestation" }
+runtime_manager_enclave = { path = "../runtime-manager", features = ["icecap"] }
+serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
+serde_json = "1.0"
+#uuid = { version = "1.1.2", features = ["serde"] }
+veracruz-utils = { path = "../veracruz-utils" }
