+++ runtime-manager/Cargo.toml
@@ -49,7 +49,6 @@ nitro = [
   "libc",
   "policy-utils/std",
   "nsm_api",
-  "nsm_lib",
   "nix",
   "raw-fd",
   "session-manager/nitro",
@@ -74,7 +73,6 @@ libm = { version = "0.2", optional = true }
 log = "0.4.13"
 nix = { version = "0.20.2", optional = true }
 nsm_api = { git = "https://github.com/aws/aws-nitro-enclaves-nsm-api.git/", tag = "v0.2.0", package = "aws-nitro-enclaves-nsm-api", optional = true }
-nsm_lib = { git = "https://github.com/aws/aws-nitro-enclaves-nsm-api.git/", tag = "v0.2.0", package = "nsm-lib", optional = true }
 policy-utils = { path = "../policy-utils" }
 protobuf = "=2.8.1"
 psa-attestation = { path = "../psa-attestation", optional = true }
