+++ runtime-manager/src/platforms/mod.rs
@@ -9,5 +9,3 @@
 //! See the `LICENSE_MIT.markdown` file in the Veracruz root directory for
 //! information on licensing and copyright.
 
-#[cfg(feature = "icecap")]
-pub mod icecap;
